docs/src/services/requirements.txt
azure.core
azure-ai-documentintelligence
streamlit
azure-storange-blob
python-dotenv